Fix unfollow error

+++ app/crud/follow.py
@@ -1,5 +1,6 @@
 from sqlalchemy.orm import Session
 from app.models.dbModel import UserFollow
+from fastapi import HTTPException
 
 def create_follow(db: Session, follower: int, followee: int) -> UserFollow:
     db_follow = db.query(UserFollow).filter(UserFollow.follow == follower, UserFollow.be_followed == followee).first()
@@ -13,7 +14,10 @@ def create_follow(db: Session, follower: int, followee: int) -> UserFollow:
 
 def delete_follow(db: Session, follower: int, followee: int) -> UserFollow:
     db_follow = db.query(UserFollow).filter(UserFollow.follow == follower, UserFollow.be_followed == followee).first()
+    if not db_follow:
+        raise HTTPException(status_code=404, detail=f"You are not following user {followee}")
     if db_follow:
         db.delete(db_follow)
         db.commit()
+    
     return db_follow
